Elastic Beanstalk - Rename and Add Implementation Coverage
