+++ moto/elasticbeanstalk/__init__.py
@@ -1,4 +1,4 @@
 from .models import eb_backends
 from moto.core.models import base_decorator
 
-mock_eb = base_decorator(eb_backends)
+mock_elasticbeanstalk = base_decorator(eb_backends)
