+++ moto/elasticbeanstalk/exceptions.py
